+++ src/app.rs
@@ -80,14 +80,16 @@ pub struct AppConfigFields {
     pub default_tree_collapse: bool,
     pub default_temp_sort_column: Option<TempWidgetColumn>,
     pub default_disk_sort_column: Option<DiskWidgetColumn>,
+    pub temperature_legend_position: Option<LegendPosition>,
 }
 
 /// For filtering out information
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Default)]
 pub struct DataFilters {
     pub disk_filter: Option<Filter>,
     pub mount_filter: Option<Filter>,
     pub temp_filter: Option<Filter>,
+    pub temp_graph_filter: Option<Filter>,
     pub net_filter: Option<Filter>,
 }
 
@@ -117,10 +119,13 @@ impl App {
         widget_map: HashMap<u64, BottomWidget>, current_widget: BottomWidget,
         used_widgets: UsedWidgets, filters: DataFilters, is_expanded: bool,
     ) -> Self {
+        let mut data_store = DataStore::new(used_widgets);
+        data_store.set_filters(filters.clone());
+
         Self {
             awaiting_second_char: false,
             second_char: None,
-            data_store: DataStore::default(),
+            data_store,
             last_key_press: Instant::now(),
             process_kill_dialog: ProcessKillDialog::default(),
             help_dialog_state: AppHelpDialogState::default(),
@@ -2038,6 +2043,31 @@ impl App {
                     }
                 }
             }
+            BottomWidgetType::TempGraph => {
+                if let Some(widget_state) = self
+                    .states
+                    .temp_graph_state
+                    .get_mut_widget_state(self.current_widget.widget_id)
+                {
+                    let new_time = widget_state
+                        .current_display_time
+                        .saturating_add(self.app_config_fields.time_interval);
+
+                    if new_time <= self.app_config_fields.retention_ms {
+                        widget_state.current_display_time = new_time;
+                        if self.app_config_fields.autohide_time {
+                            widget_state.autohide_timer = Some(Instant::now());
+                        }
+                    } else if widget_state.current_display_time
+                        != self.app_config_fields.retention_ms
+                    {
+                        widget_state.current_display_time = self.app_config_fields.retention_ms;
+                        if self.app_config_fields.autohide_time {
+                            widget_state.autohide_timer = Some(Instant::now());
+                        }
+                    }
+                }
+            }
             _ => {}
         }
     }
@@ -2116,6 +2146,29 @@ impl App {
                     }
                 }
             }
+            BottomWidgetType::TempGraph => {
+                if let Some(widget_state) = self
+                    .states
+                    .temp_graph_state
+                    .get_mut_widget_state(self.current_widget.widget_id)
+                {
+                    let new_time = widget_state
+                        .current_display_time
+                        .saturating_sub(self.app_config_fields.time_interval);
+
+                    if new_time >= STALE_MIN_MILLISECONDS {
+                        widget_state.current_display_time = new_time;
+                        if self.app_config_fields.autohide_time {
+                            widget_state.autohide_timer = Some(Instant::now());
+                        }
+                    } else if widget_state.current_display_time != STALE_MIN_MILLISECONDS {
+                        widget_state.current_display_time = STALE_MIN_MILLISECONDS;
+                        if self.app_config_fields.autohide_time {
+                            widget_state.autohide_timer = Some(Instant::now());
+                        }
+                    }
+                }
+            }
             _ => {}
         }
     }
@@ -2162,11 +2215,25 @@ impl App {
         }
     }
 
+    fn reset_temp_graph_zoom(&mut self) {
+        if let Some(widget_state) = self
+            .states
+            .temp_graph_state
+            .get_mut_widget_state(self.current_widget.widget_id)
+        {
+            widget_state.current_display_time = self.app_config_fields.default_time_value;
+            if self.app_config_fields.autohide_time {
+                widget_state.autohide_timer = Some(Instant::now());
+            }
+        }
+    }
+
     fn reset_zoom(&mut self) {
         match self.current_widget.widget_type {
             BottomWidgetType::Cpu => self.reset_cpu_zoom(),
             BottomWidgetType::Mem => self.reset_mem_zoom(),
             BottomWidgetType::Net => self.reset_net_zoom(),
+            BottomWidgetType::TempGraph => self.reset_temp_graph_zoom(),
             _ => {}
         }
     }