+++ src/canvas.rs
@@ -266,6 +266,12 @@ impl Painter {
                         #[cfg(feature = "battery")]
                         self.draw_battery(f, app_state, rect[0], app_state.current_widget.widget_id)
                     }
+                    TempGraph => self.draw_temperature_graph(
+                        f,
+                        app_state,
+                        rect[0],
+                        app_state.current_widget.widget_id,
+                    ),
                     _ => {}
                 }
             } else if app_state.app_config_fields.use_basic_mode {
@@ -380,6 +386,12 @@ impl Painter {
                                 #[cfg(feature = "battery")]
                                 self.draw_battery(f, app_state, vertical_chunks[3], widget_id)
                             }
+                            TempGraph => self.draw_temperature_graph(
+                                f,
+                                app_state,
+                                vertical_chunks[3],
+                                widget_id,
+                            ),
                             _ => {}
                         }
                     }
@@ -457,6 +469,9 @@ impl Painter {
                     {
                         #[cfg(feature = "battery")]
                         self.draw_battery(f, app_state, *draw_loc, widget.widget_id)
+                    }
+                    TempGraph => {
+                        self.draw_temperature_graph(f, app_state, *draw_loc, widget.widget_id)
                     }
                     _ => {}
                 }