+++ src/widgets/network_graph.rs
@@ -1,15 +1,11 @@
 use std::time::Instant;
 
+use crate::widgets::GraphHeightCache;
+
 pub struct NetWidgetState {
     pub current_display_time: u64,
     pub autohide_timer: Option<Instant>,
-    pub height_cache: Option<NetWidgetHeightCache>,
-}
-
-pub struct NetWidgetHeightCache {
-    pub best_point: (Instant, f64),
-    pub right_edge: Instant,
-    pub period: u64,
+    pub height_cache: GraphHeightCache,
 }
 
 impl NetWidgetState {
@@ -17,7 +13,7 @@ impl NetWidgetState {
         NetWidgetState {
             current_display_time,
             autohide_timer,
-            height_cache: None,
+            height_cache: GraphHeightCache::default(),
         }
     }
 }